chore: update tags page
